Added utils app, and makestaticdirs command.
The utils app is meant to store anything that will help me with development.
For now, that's just a number of manage.py commands.

The makestatic dirs command will create the dirs for storing static content.
The makestatic dirs command is not complete. All it does now is check that
an app exists and print the dirs that it should create. I'll polish the
command off later today.

I will also add a checkstaticdirs command that makes sure the dirs are
organized properly.