First version of the hero.
DONE
Used Bulma's is-hidden (display: none;) to hide the words for the acronym.
Created a js function that will show the words when the corresponding letters
are hovered in KGI STEAM.
Ensured that hero works well on mobile and desktop.

TODO
Clarify html conventions.
Reqrite the js functions for reusability and readability.