Mathematical expressions as valid worksheet answers.
DONE
Users can enter a mathematical expression into the form
on the worksheets and the computer will caluculate the
appropriate value.

TODO
If the user enters an invalid expression. The form raises a
validation error. However, the reidrect is not allowing the
error to propogate and the end user never sees the error message.
Perhaps it's time to untangle the CourseView class....