Helpful message if .env file is missing.
DONE
====
I ran into some trouble getting the system running on my local
machine after I deleted the git repo. The issue came from the
fact that I forgot what was in the .env_settings file.

manage.py will throw an error when the file is missing and
give the user direction on how to fix it.